backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-mind-orchestrator"
version = "0.1.0"
description = "CLI-first, security-first LLM orchestrator"
requires-python = ">=3.11"
dependencies = [
  "argon2-cffi>=23.1.0",
  "anthropic>=0.60.0",
  "click>=8.1.7",
  "cryptography>=43.0.0",
  "fastapi>=0.115.0",
  "google-genai>=0.3.0",
  "httpx>=0.27.0",
  "keyring>=25.0.0",
  "openai>=1.51.0",
  "pyyaml>=6.0.2",
  "rich>=13.8.1",
  "tiktoken>=0.8.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.24.0",
]
server = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
]
discord = [
  "discord.py>=2.4.0",
]

[project.scripts]
mmctl = "mmctl.__main__:main"

[tool.setuptools]
packages = ["mmctl", "orchestrator", "orchestrator.providers", "orchestrator.collaboration", "orchestrator.security", "orchestrator.observability", "orchestrator.tools", "orchestrator.retrieval", "orchestrator.memory", "orchestrator.prompts", "orchestrator.skills", "evaluation", "integrations"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
